// embedded/src/types/database.ts
export interface Table {
  id: string;
  base_id: string;
  title: string;
  description: string;
  position: number;
  created_at: string;
  updated_at: string;
}
// ...
export interface CreateTableInput {
  base_id: string;
  title: string;
  description?: string;
}

export interface UpdateTableInput {
  title?: string;
  description?: string;
}